requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
datasets==3.5.0
dill==0.3.8
docker-pycreds==0.4.0
filelock==3.18.0
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.4
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.8.0.87
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
# nvidia-nccl-cu12==2.26.2
nvidia-nccl-cu12==2.25.1
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.8.55
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.7
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
# pytorch-triton==3.3.0+git96316ce5
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
sentry-sdk==2.25.1
setproctitle==1.3.5
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sympy==1.13.3
# torch==2.8.0.dev20250428+cu128
# torchaudio==2.6.0.dev20250428+cu128
# torchvision==0.22.0.dev20250428+cu128
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wandb==0.19.9
wheel==0.45.1
xxhash==3.5.0
yarl==1.19.0
mamba-ssm==2.2.4
lion-pytorch==0.2.3